+++ src/daemon/handler/srv_data.cpp
@@ -114,9 +114,10 @@ rpc_srv_write(hg_handle_t handle) {
             "{}() path: '{}' chunk_start '{}' chunk_end '{}' chunk_n '{}' total_chunk_size '{}' bulk_size: '{}' offset: '{}'",
             __func__, in.path, in.chunk_start, in.chunk_end, in.chunk_n,
             in.total_chunk_size, bulk_size, in.offset);
-
-    GKFS_DATA->stats()->add_value_size(gkfs::utils::Stats::SizeOp::write_size,
-                                       bulk_size);
+    if(GKFS_DATA->enable_stats()) {
+        GKFS_DATA->stats()->add_value_size(
+                gkfs::utils::Stats::SizeOp::write_size, bulk_size);
+    }
 
 #ifdef GKFS_ENABLE_AGIOS
     int* data;
@@ -238,9 +239,10 @@ rpc_srv_write(hg_handle_t handle) {
                     __func__, chnk_id_file, host_id, chnk_id_curr);
             continue;
         }
-#ifdef GKFS_CHUNK_STATS
-        GKFS_DATA->stats()->add_write(in.path, chnk_id_file);
-#endif
+
+        if(GKFS_DATA->enable_chunkstats()) {
+            GKFS_DATA->stats()->add_write(in.path, chnk_id_file);
+        }
 #endif
 
         chnk_ids_host[chnk_id_curr] =
@@ -412,9 +414,10 @@ rpc_srv_read(hg_handle_t handle) {
             "{}() path: '{}' chunk_start '{}' chunk_end '{}' chunk_n '{}' total_chunk_size '{}' bulk_size: '{}' offset: '{}'",
             __func__, in.path, in.chunk_start, in.chunk_end, in.chunk_n,
             in.total_chunk_size, bulk_size, in.offset);
-
-    GKFS_DATA->stats()->add_value_size(gkfs::utils::Stats::SizeOp::read_size,
-                                       bulk_size);
+    if(GKFS_DATA->enable_stats()) {
+        GKFS_DATA->stats()->add_value_size(
+                gkfs::utils::Stats::SizeOp::read_size, bulk_size);
+    }
 
 #ifdef GKFS_ENABLE_AGIOS
     int* data;
@@ -525,7 +528,9 @@ rpc_srv_read(hg_handle_t handle) {
                     __func__, chnk_id_file, host_id, chnk_id_curr);
             continue;
         }
-        GKFS_DATA->stats()->add_read(in.path, chnk_id_file);
+        if(GKFS_DATA->enable_chunkstats()) {
+            GKFS_DATA->stats()->add_read(in.path, chnk_id_file);
+        }
 #endif
 
         chnk_ids_host[chnk_id_curr] =