+++ tests/integration/harness/gkfs.py
@@ -252,7 +252,9 @@ class Daemon:
                  '-l', self._address,
                  '--metadir', self._metadir,
                  '--dbbackend', self._database, 
-                 '--output-stats', self.logdir / 'stats.log' ]
+                 '--output-stats', self.logdir / 'stats.log',
+                 '--enable-collection',
+                 '--enable-chunkstats' ]
         if self._database == "parallaxdb" :
             args.append('--clean-rootdir-finish')